pyproject.toml
[project]
name = "alma-letters"
version = "0.1.0"
description = "Export customized Alma letter templates via the Configuration API"
requires-python = ">=3.9"
dependencies = [
    "requests",
    "python-dotenv",
]

[project.scripts]
alma-letters = "export_letters:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["export_letters"]
